Updates based on code review
The biggest change here is that we are changing the data model, moving
the flag field from the Confirmation record to the Term record. This
clears the way for Confirmation to treat Category as a required field,
which explains the three migrations in this commit.

Because "flagged" is now a category in Confirmation terms, we add that
to the seeds file.

Along with the data model change, we update the new confirmation form to
more explicitly label fields as ID fields, in addition to no longer
having a separate list entry for flagging terms.

Through all of the above, the controller shifts a bit, as set_category
is no longer needed to manage received values. In its place, there is
now a flag_term method for setting the new flag on the related term - as
long as the confirmation saves correctly.

There is also a guard clause for when the confirmation does not save
correctly, which means the feedback_for method gets a little simpler.

---
Next up is a slight refator to adopt a strong parameters approach,
which will have some additional changes with it.