Remove Fingerprint calculation from migration
This comments out the initial calculation of Fingerprint records from
the migrations in this PR. A following commit will add a rake task that
will focus on this work.

A side effect of this work is that we need to change the fixtures used
for our test suite, to match the state the migrations will produce. This
in turn causes some tests to change, because some tests will require
additional setup work.

More complex tests now have code comments to indicate this sort of
sequencing.