Refactor Fingerprint model to use "value" field
Until now, we've been building the Fingerprint model to include a field
also named "fingerprint" for the actual string value of the fingerprint.
This removes that duplication by changing the field name to be "value".

As a result of updating the migration with this new terminology, we also
update the alias / delegation up to the Term model, so terms can still
find the string via the term.fingerprint_value attribute.

The fixtures use the updated field name, and some references to the new
field name (in application code and in tests) are updated.