Add Journal model for records management
** Why are these changes being introduced:

Our current Detector::Journal model is asked to do two separate tasks:
1. Detect whether incoming search terms match with a list of known
   journals
2. Maintain that list of known journals

This conflation of concerns prevents the model from using some of the
shared tooling we are developing, like the Bulk Checker module, because
the Bulk Checker ends up calling the .new() method (which doesn't look
for a new detection, but creates a new journal record).

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-110

** How does this address that need:

This commit takes the first step toward resolving this conflict, by
creating a new Journal model that will ake over the management of
journal records. The Detector::Journal will then be simplified in a
future commit to only deal with the detection operations.

Specifically:
1. Adds a migration to create the Journal database table, matching the
structure of the Detector::Journal table.

2. Creates test fixtures for these records

3. Updates the existing rake tasks to manage Journal records, and not
   Detector::Journal.

4. The new ::Journal model copies the records-management methods and
   lifecycle hook from Detector::Journal (downcasing journal title)

5. The existing Detector::Journal model has its _match methods updated
   to look up matches using the ::Journal model - which is initially
   unpopulated (see side effects)

6. Tests are updated to reflect the use of the new Journal model.

** Document any side effects to this change:

As of this commit, there is a bit of duplication in the application,
which will be resolved in future commits in this PR.

This commit also does not initially populate the new Journal table,
so the updated rake task will need to be run as part of the promotion
of this work. During the period of time between the deployment and
running this rake task, the application will fail to detect journal
matches correctly.