Convert Detector::Journal to plain ruby object
** Why are these changes being introduced:

The previous commit on this branch created a new Journal model and table
for the management of journal records, but did not remove that
functionality from the existing Detector::Journal model.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-110

** How does this address that need:

This finishes the job, removing the database table for detector_journals
and converting the Detector::Journal model from an ActiveRecord type to
a "plain old ruby object" (PORO).

During this change, we also remove the test fixtures, and the one test
of this model that confirmed this records management function. That test
was previously moved to the Journal test suite.

** Document any side effects to this change:

None