Add Bulk Checker module to Detector::Journal
** Why are these changes being introduced:

Now that the Detector::Journal model is only concerns with detections,
and calling Detector::Journal.new no longer spawns any database records,
we can implement the BulkChecker module for better developer experience.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-110

** How does this address that need:

This defines a @detections attribute on the Detector::Journal class,
as well as an initialize method to populate that attribute based on the
full_term_match method.

This allows the BulkChecker-provided .check_all_matches method to return
meaningful information.

** Document any side effects to this change:

This implementation may be too minimal, as the initialize method ends
up calling the singleton methods in a slightly awkward way (or maybe I
need to get used to this way of working, and not expect to just call
full_term_matches directly in the initializer).