Update test fixtures and seeds
** Why are these changes being introduced:

We are going to add a new Detector that looks for MIT barcodes, with the
ability to look up the items using the Primo API. This will involve
creating a small number of database records which correspond to the
class coming in the next commit.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-72

** How does this address that need:

This updates our test fixtures and database seeds with the records that
anticipate the coming detector class. No functionality is changed in the
application with this commit, but the tests all still pass.

There is a Detector record for the barcode detector, and a mapping
record to link the detector with a Transactional category. I've set the
confidence pretty high, as the format of the barcode is highly distinct.

** Document any side effects to this change:

None