Add barcode regex to standard identifier
** Why are these changes being introduced:

Barcodes are detected as part of the StandardIdentifier class, using a
pretty simple regex.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-72

** How does this address that need:

This adds the regex to detect a barcode as a standalone search (not as
part of a longer search string).

** Document any side effects to this change:

Adding this regex pattern, with the fixtures added in the last commit,
causes one test to need an updated expected record count.