Add Barcode lookup model
** Why are these changes being introduced:

The biggest part of this ticket is to build a lookup class which takes
the detected barcode and looks it up in the Primo API. Otherwise TACOS
would just say "yep, that's a barcode alright" and do nothing more.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-72

** How does this address that need:

This adds the lookup class for dealing with the Primo API, using the XML
format which Primo returns (other options are available, but not JSON).

As much as possible I've tried to use similar structure to the other
lookup classes, particularly the PubMed lookup which also deals with XML
format.

The StandardIdentifiers record type in GraphQL is also expanded to call
this lookup, and to acknowledge that it exists in the docs. See the side
effects below for a note about this.

For tests, we add some model tests that flex the lookup class, as well
as a controller test that confirms the lookup is called correctly. Along
the way we add two VCR cassettes for the API interactions.

** Document any side effects to this change:

The GraphQL controller now trips two code quality metrics, for length
and for complexity. The length is 12 lines (more than 10), and the
branching complexity is above the 18 threshold. Given that the entire
method is a case block, I disagree that this is too complex for this PR.
There _is_ an alternative structure that I've seen used, but I'd rather
talk about that separately.

The lookup class strips out the namespaces from the XML response. I'm
not happy about this, but it was the shortest path to finishing this
and I don't think we're likely to deal with namespace collisions in
this API as a result of stripping them out.

I'm not entirely happy with using a fulldisplay URL in the name of a
"link resolver URL" in the lookup class, but the complexity of lookups
returning multiple URL types feels like a separate concern than this
ticket.