Integrate barcodes with metrics report
** Why are these changes being introduced:

The barcode detections need to be included in the monthly and cumulative
metrics reports we generate.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-72

** How does this address that need:

This adds a barcode field to the metrics table, and updates the model to
include these totals. Tests in the model are also updated, along with
some new fixture records for search events that will trigger everything.

** Document any side effects to this change:

No