Updates based on code review feedback
- The name of a variable in the Term model test suite has been updated
  to hopefully be clearer about its use.

- The confirmation workflow is now restrictd to already-categorized
  terms for non-admin users. Admins will still have an ability to load
  all terms for this workflow, but as noted this is not the smoothest
  workflow yet.

Side note: the restriction being enacted here is on the ability to load
the _index_ in a specific state, but the actual confirmation form is not
affected. A non-admin user who understands the URL structure will be
able to guess the right URL for a given term (or pick one randomly), and
submit a confirmation for a term that has not been categorized.

I believe this is an acceptable risk, as the impact of such a user is
that we'd have some extra confirmation records in the system - which are
useful only for adjusting future operations, so I struggle to think of
a downside to having some extras.