Add label field to Term model
This allows the application to support labelled records for the evaluation of its internal algorithms.

Along with this change, the rake task to load search events - which is how labelled data will be supplied to the application - is updated to support a label field.

This includes a new test to confirm that support, and a new cassette holding labelled data.