Implement lookup_citation model to deal w/ lambda
** Why are these changes being introduced:

We need a way for TACOS to communicate with the new detector lambda that
serves the next iteration of our citation detector.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-183

** How does this address that need:

* The big change is that we add a lookup_citation model that handles
communication with the lambda. The patterns used are similar to our
other external integrations (the other lookup_ classes) - although this
is not consistently implemented.

* A secondary change is to define a fourth instance variable on the
citation model, for @features. This is compiled from the existing work,
reformatting and joining two other instance variables.

* There is a mismatch between the features generated by the existing
citation class and those expected by the new detector. Two features
have different names, while one feature is not used at all. That change
is handled via the extract_features method on lookup_citation, in order
to make these changes as close as possible to the external call.

** Document any side effects to this change:

* Using a lookup_* pattern for this is a departure from the way the
other classes are used, and I'm tempted to refactor this PR in order
to adopt a gateway pattern instead. The other lookup classes are used
by GraphQL to look up external information after a detector has fired -
while this lookup_citation is much earlier in the application flow,
being part of the detector itself.