Add @features to citation detector
** Why are these changes being introduced:

In order to integrate TACOS with the external lambda, we need to send a
hash of the extracted features. This is done implicitly within the
existing citation detector, but not exposed in the format the lambda
will need.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-183

** How does this address that need:

This adds a fourth instance variable to the citation detector: @features
This variable is in the format needed by the lambda, and is derived from
the existing variables @subpatterns and @summary.

There is a new private method which handles this compilation, which does
a semi-complex bit of manipulation. I've tried to comment its operation
thoroughly.

** Document any side effects to this change:

The instance variables provided by the citation detector now have some
duplication, as everything in the @summary variable is also contained
in the @features variable. Ultimately we'll probably be able to remove
the older instance variables, but that will be a separate refactor.