Include mlcitation detector in term operations
** Why are these changes being introduced:

The term model will need to interact with the new detector, but only
when we are ready to use it.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-183

** How does this address that need:

This adds the call to Detector::MlCitation.record(term) in the term
model's record_detections, model, but guarded by a check of the
environment variables.

As I write this, the detector has not been deployed to AWS, so that
guard is relied upon heavily. It is also why these three env vars have
not been defined anywhere in Heroku.

There is one new test on the term model confirming both sides of this
guard clause.

** Document any side effects to this change:

None, I don't think.