Adds new detector::ml_citation model
** Why are these changes being introduced:

We need a detector class that will handle communication with the new
external machine learning algorithm for citation detection.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-183

** How does this address that need:

This creates a new detector, detector::ml_citation. Along with this
class, we create seeds and fixtures which treat this as a _separate_
detector from the existing citation detector (see side effects).

As of this commit, the detector is not integrated with the rest of the
application, as the term model's record_detections method has not yet
been changed.

There are also tests, along with three VCR cassettes which simulate its
responses in the test suite. Please note, though, that the detector
requires a trio of environment variables to be set in order to do any
work. Those variables have not been set in .env.test, because that will
cause a large number of tests that have nothing to do with this feature
to now need cassettes. Instead, there is a helper function that needs
to be called for any test that needs the new ML citation detector.

For local development, you'll want to have the running lambda on your
laptop somewhere (see the tacos-dtectors-lambdas repo for instructions).

** Document any side effects to this change:

I'm not sure we will want this to be a separate / parellel detector, or
if this should be a replacement for the existing citation detector while
updating the app's DETECTOR_VERSION variable. This sort of significant
upgrade to functionality is the sort of use case we envisioned for that
variable in the first place, because having this new capability should
mean that we re-process terms that had already been evaluated and
categorized.