tco-190 Filter searches before citation detector
** Why are these changes being introduced:

There is a transaction cost to calling the citation detector, both in
terms of time and processing time. If we can develop a way to flag
searches which have zero chance of being a citation, we can cut out this
expense by skipping the citation detector.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-190

** How does this address that need:

This moves the extract_features method in Detector::MlCitation into the
initialize method, allowing us to quickly check for whether a given term
has enough non-zero features to make it worth calling the detector.

From our analysis in the TACOS notebooks, we believe that phrases which
result in only two non-zero values among all their features will never
end up being a citation - and that this threshold will allow us to skip
the citation detector in 90% of searches.

The filtering is performed in a convenience method named
enough_nonzero_values? (naming things is hard).

** Document any side effects to this change:

The @detections instance variable is defined as false at the top of
the initialize method, before the first guard clause, so that we get a
consistent Boolean value in all conditions. This required one test to
change that previously expected a nil.