Remove lastnames feature from MlCitation payload
** Why are these changes being introduced:

* We are removing the lastnames feature from the MlCitation detector.
It will still be extracted by Detector::Citation, but the algorithm
within the detector won't support it.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-155

** How does this address that need:

This filters out the lastnames feature from Detector::MlCitation's
extract_features method, which prevents it from being sent to the
lambda.

The cassettes are also regenerated. Two of them are just removed, which
confuses me as they should still be needed.

** Document any side effects to this change:

None, as far as I know